server/mcp_server_veen/python/pyproject.toml
[project]
name = "mcp-server-veen"
version = "0.1.0"
description = "Apply, configure, and query for edge computing nodes, including virtual machines, images, bare metal, and corresponding network configurations."
readme = "../README.md"
requires-python = ">=3.11"
dependencies = [
    "mcp>=1.6.0",
    "pydantic>=2.11.3",
    "python-dotenv>=1.1.0",
    "retry>=0.9.2",
    "starlette>=0.46.2",
    "volcengine>=1.0.180",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]